// src/app/interfaces/condicion5/condicion5.ts
export interface Condicion5 {
    aspecto1Programa: string;
    aspecto1File: string;
    aspecto2Programa: string;
    aspecto2File: string;
    aspecto3Programa: string;
    aspecto3File: string;
    aspecto4Programa: string;
    aspecto4File: string;
    aspecto5Programa: string;
    aspecto5File: string;
    aspecto6Programa: string;
    aspecto6File: string;
    key$?: string;
}